# Iri parsing performance improvement