--- a/modules/core/src/main/scala/ch/epfl/bluebrain/nexus/rdf/Curie.scala
+++ b/modules/core/src/main/scala/ch/epfl/bluebrain/nexus/rdf/Curie.scala
@@ -1,12 +1,9 @@
 package ch.epfl.bluebrain.nexus.rdf
 
-import cats.syntax.either._
 import cats.syntax.show._
 import cats.{Eq, Show}
 import ch.epfl.bluebrain.nexus.rdf.Curie.Prefix
 import ch.epfl.bluebrain.nexus.rdf.Iri.RelativeIri
-import org.parboiled2.ErrorFormatter
-import org.parboiled2.Parser.DeliveryScheme.{Either => E}
 
 /**
   * A Compact URI as defined by W3C in ''CURIE Syntax 1.0''.
@@ -28,9 +25,7 @@ object Curie {
     * @return Right(Curie) if the string conforms to specification, Left(error) otherwise
     */
   final def apply(string: String): Either[String, Curie] =
-    new IriParser(string).`curie`
-      .run()
-      .leftMap(_.format(string, formatter))
+    new IriParser(string).parseCurie
 
   /**
     * The Compact URI prefix as defined by W3C in ''CURIE Syntax 1.0''.
@@ -49,9 +44,7 @@ object Curie {
       * @return Right(prefix) if the string conforms to specification, Left(error) otherwise
       */
     final def apply(string: String): Either[String, Prefix] =
-      new IriParser(string).`prefix`
-        .run()
-        .leftMap(_.format(string, formatter))
+      new IriParser(string).parseNcName
 
     final implicit val prefixShow: Show[Prefix] = Show.show(_.value)
     final implicit val prefixEq: Eq[Prefix]     = Eq.fromUniversalEquals
@@ -61,6 +54,4 @@ object Curie {
     Show.show { case Curie(prefix, reference) => prefix.show + ":" + reference.show }
 
   final implicit val curieEq: Eq[Curie] = Eq.fromUniversalEquals
-
-  private val formatter = new ErrorFormatter(showExpected = false, showTraces = false)
 }