--- a/modules/core/src/test/scala/ch/epfl/bluebrain/nexus/rdf/UrnSpec.scala
+++ b/modules/core/src/test/scala/ch/epfl/bluebrain/nexus/rdf/UrnSpec.scala
@@ -12,14 +12,13 @@ class UrnSpec extends WordSpecLike with Matchers with Inspectors with EitherValu
       // format: off
       val cases = List(
         "urn:uUid:6e8bc430-9c3a-11d9-9669-0800200c9a66"           -> "urn:uuid:6e8bc430-9c3a-11d9-9669-0800200c9a66",
-        "urn:example:a%C2%A3/b%C3%86c//:://?=a=b#"                 -> "urn:example:a£/bÆc//:://?=a=b#",
+        "urn:example:a%C2%A3/b%C3%86c//:://?=a=b#"                -> "urn:example:a£/bÆc//:://?=a=b#",
         "urn:lex:eu:council:directive:2010-03-09;2010-19-UE"      -> "urn:lex:eu:council:directive:2010-03-09;2010-19-UE",
         "urn:Example:weather?=op=map&lat=39.56&lon=-104.85#test"  -> "urn:example:weather?=lat=39.56&lon=-104.85&op=map#test",
         "urn:examp-lE:foo-bar-baz-qux?+CCResolve:cc=uk"           -> "urn:examp-le:foo-bar-baz-qux?+CCResolve:cc=uk",
         "urn:examp-lE:foo-bar-baz-qux?=a=b?+CCResolve:cc=uk"      -> "urn:examp-le:foo-bar-baz-qux?+CCResolve:cc=uk?=a=b",
         "urn:examp-lE:foo-bar-baz-qux?+CCResolve:cc=uk?=a=b"      -> "urn:examp-le:foo-bar-baz-qux?+CCResolve:cc=uk?=a=b",
         "urn:examp-lE:foo-bar-baz-qux?+CCResolve:cc=uk?=a=b#hash" -> "urn:examp-le:foo-bar-baz-qux?+CCResolve:cc=uk?=a=b#hash"
-
       )
       // format: on
       forAll(cases) {
@@ -29,7 +28,10 @@ class UrnSpec extends WordSpecLike with Matchers with Inspectors with EitherValu
     }
 
     "fail to parse" in {
-      val fail = List("urn:example:some/path/?=")
+      val fail = List(
+        "urn:example:some/path/?+",
+        "urn:example:some/path/?=",
+      )
       forAll(fail) { str =>
         Urn(str).left.value
       }