… bench/src/main/resources/schema.json